Automatic Login
On this tutorial, it's explained of how to login on a website automatically. It only works for simple logins, without captchas or any other steps that detect bots or send codes to your phone. 
It's an interesting feature that can be associated to send an automatic e-mail, as long as your account is open on the browser.